backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
botbuilder-core>=4.16.2
botbuilder-integration-aiohttp>=4.16.2
botframework-connector>=1.5.0
slack_sdk>=3.27.0
